@@ -9,12 +9,12 @@ var Commander = require('./commander');
 var utils = require('./utils');
 var asCallback = require('standard-as-callback').default;
 var eventHandler = require('./redis/event_handler');
-var debug = require('./utils/debug')('ioredis:redis');
 var {StandaloneConnector, SentinelConnector} = require('./connectors');
 var ScanStream = require('./ScanStream').default;
 var commands = require('redis-commands');
 var PromiseContainer = require('./promiseContainer');
 
+var debug = utils.Debug('redis')
 /**
  * Creates a Redis instance
  *
@@ -296,7 +296,6 @@ Redis.prototype.connect = function (callback) {
       stream.once(CONNECT_EVENT, eventHandler.connectHandler(_this));
       stream.once('error', eventHandler.errorHandler(_this));
       stream.once('close', eventHandler.closeHandler(_this));
-      stream.on('data', eventHandler.dataHandler(_this));
 
       if (options.connectTimeout) {
         /*
@@ -394,6 +393,40 @@ Redis.prototype.end = function () {
 Redis.prototype.duplicate = function (override) {
   return new Redis(Object.assign({}, this.options, override || {}));
 };
+
+Redis.prototype.recoverFromFatalError = function (commandError, err, options) {
+  this.flushQueue(err, options);
+  this.silentEmit('error', err);
+  this.disconnect(true);
+}
+
+Redis.prototype.handleReconnection = function (err, item) {
+  var needReconnect = false;
+  if (this.options.reconnectOnError) {
+    needReconnect = this.options.reconnectOnError(err);
+  }
+
+  switch (needReconnect) {
+  case 1:
+  case true:
+    if (this.status !== 'reconnecting') {
+      this.disconnect(true);
+    }
+    item.command.reject(err);
+    break;
+  case 2:
+    if (this.status !== 'reconnecting') {
+      this.disconnect(true);
+    }
+    if (this.condition.select !== item.select && item.command.name !== 'select') {
+      this.select(item.select);
+    }
+    this.sendCommand(item.command);
+    break;
+  default:
+    item.command.reject(err);
+  }
+}
 
 /**
  * Flush offline queue and command queue with error.
@@ -695,7 +728,5 @@ Redis.prototype._getDescription = function () {
     }, options));
   };
 });
-
-Object.assign(Redis.prototype, require('./redis/parser'));
 
 module.exports = Redis;
